refactor: rename nautobot/e2e to nautobot/playwright
The shared infrastructure is Playwright-specific, so it is named for
the tool; the test directories keep scope-based names. Verified that
nautobot.playwright does not shadow the playwright package for
absolute imports.
