+++ nautobot/core/tests/integration/__init__.py
@@ -1,3 +1,3 @@
 """App-agnostic E2E machinery tests (the framework itself, not a product page)."""
 
-from nautobot.e2e import load_tests  # noqa: F401  keeps unittest discovery out; see nautobot.e2e
+from nautobot.playwright import load_tests  # noqa: F401  keeps unittest discovery out; see nautobot.playwright
