+++ nautobot/dcim/tests/integration/__init__.py
@@ -1,3 +1,3 @@
 """Playwright E2E tests for the DCIM app. Run with `invoke e2e --app dcim`."""
 
-from nautobot.e2e import load_tests  # noqa: F401  keeps unittest discovery out; see nautobot.e2e
+from nautobot.playwright import load_tests  # noqa: F401  keeps unittest discovery out; see nautobot.playwright
