+++ nautobot/dcim/tests/integration/conftest.py
@@ -1,15 +1,15 @@
 """DCIM E2E fixtures: thin named fixtures over the shared `create_object` factory.
 
-The factory (`nautobot.e2e.fixtures`) owns creation and teardown;
+The factory (`nautobot.playwright.fixtures`) owns creation and teardown;
 fixtures here decide which objects a DCIM test starts from.
 Shared fixtures (create_object, status_id_for, auth_page, ...) are provided by
-nautobot.e2e.fixtures, registered in the repo-root conftest; run
+nautobot.playwright.fixtures, registered in the repo-root conftest; run
 `pytest --fixtures` to list them.
 """
 
 import pytest
 
-from nautobot.e2e.fixtures import unique_name
+from nautobot.playwright.fixtures import unique_name
 
 
 @pytest.fixture
